refactor : package refactoring - board package로 이동
