+++ src/main/kotlin/next/step/minesweeper/domain/board/BoardHeight.kt
@@ -1,7 +1,8 @@
-package next.step.minesweeper.domain
+package next.step.minesweeper.domain.board
 
 @JvmInline
 value class BoardHeight(val height: Int) {
+
     init {
         require(height > MIN_HEIGHT) { "높이는 ${MIN_HEIGHT}보다 커야합니다." }
     }
