+++ src/main/kotlin/next/step/minesweeper/domain/board/BoardWidth.kt
@@ -1,7 +1,8 @@
-package next.step.minesweeper.domain
+package next.step.minesweeper.domain.board
 
 @JvmInline
 value class BoardWidth(val width: Int) {
+
     init {
         require(width > MIN_WIDTH) { "너비는 ${MIN_WIDTH}보다 커야합니다." }
     }
