+++ src/test/kotlin/next/step/minesweeper/domain/board/BoardHeightTest.kt
@@ -1,18 +1,18 @@
-package next.step.minesweeper.domain
+package next.step.minesweeper.domain.board
 
 import io.kotest.core.spec.style.DescribeSpec
 import io.kotest.datatest.withData
 import org.junit.jupiter.api.assertThrows
 
-class BoardWidthTest : DescribeSpec({
-    describe("BoardWidth") {
+class BoardHeightTest : DescribeSpec({
+    
+    describe("BoardHeight") {
         context("0이하의 값으로 생성하면 예외 발생") {
             withData(
                 listOf(-1, 0)
-            ) { width ->
-                assertThrows<IllegalArgumentException> { BoardWidth.of(width) }
+            ) { height ->
+                assertThrows<IllegalArgumentException> { BoardHeight.of(height) }
             }
         }
     }
-
 })
